+++ guillotina/tests/test_dbobjects.py
@@ -50,14 +50,14 @@ async def test_create_annotation(dummy_txn_root):
         assert ob2._p_jar is None
         assert ob2.__name__ is None
         assert ob2.__parent__ is None
-        assert len(ob1.__annotations__) == 0
+        assert len(ob1.__gannotations__) == 0
 
         await annotations.async_set('test2', ob2)
         assert ob2.__of__ is ob1._p_oid
         assert ob2._p_jar is ob1._p_jar
         assert ob2.__name__ == 'test2'
         assert ob2.__parent__ is None
-        assert len(ob1.__annotations__) == 1
+        assert len(ob1.__gannotations__) == 1
 
 
 async def test_use_behavior_annotation(dummy_txn_root):