+++ CHANGELOG.rst
@@ -4,6 +4,10 @@
 - Implement minimal passing mypy compatibility
   [vangheem]
 
+- Rename `BaseObject.__annotations__` to `BaseObject.__gannotations__` to prevent
+  namespace clashes with mypy and other things
+  [vangheem]
+
 
 4.0.0 (2018-06-05)
 ------------------